phonecalls/requirements.txt
Django==2.0.5
djangorestframework==3.8.2
python-dateutil==2.7.3
coreapi==2.3.3
